feat: 確認ページの分離
